Refactor: Combined Models Folders
