+++ app/src/main/java/com/app/venki/up/model/coupons/CouponsArray.java
@@ -1,4 +1,4 @@
-package com.app.venki.up.models.events.coupons;
+package com.app.venki.up.model.coupons;
 
 import java.util.List;
