+++ app/src/main/java/com/app/venki/up/recycleradapters/GroupOnRecycler.java
@@ -14,7 +14,7 @@ import android.widget.TextView;
 
 import com.bumptech.glide.Glide;
 import com.app.venki.up.R;
-import com.app.venki.up.models.events.coupons.GroupOnEvent;
+import com.app.venki.up.model.coupons.GroupOnEvent;
 
 
 public class GroupOnRecycler extends android.support.v7.widget.RecyclerView.Adapter<GroupOnRecycler.RecyclerViewHolder> {
